Do not error when closing unhandled component
Ensure that `close` can be called on the stripe service even if a
StripeCheckout handler had not yet been configured for that component.

Since the stripe handler is registered via `_stripeHandler`, a component that
is destroyed before the StripeCheckout JS is loaded would produce a
`StripeCheckout is undefined` error when invoking `_stripeHandler`.

One solution to this would have been to wrap the `close` function in the same
promise `open` uses to ensure StripeCheckout was available, but it seems to me
that `close` shouldn’t be responsible for registering a new handler simply to
delete it.

This also fixes previously broken tests.